Move Player-related functions to a new Player module
This is done to make a clearer separation of concerns in the game, allowing the
GameState to only really care about the state of the game.
