+++ lib/socketfight/game_state.ex
@@ -4,6 +4,7 @@
 defmodule Socketfight.GameState do
   use Agent
   alias Socketfight.CollisionDetector
+  alias Socketfight.Player
 
   # @arena_width 1080
   # @arena_height 720
@@ -73,7 +74,7 @@ defmodule Socketfight.GameState do
       player ->
         player
         |> update_player_meta(action, state)
-        |> update_player
+        |> update_player()
     end
   end
 
@@ -88,20 +89,13 @@ defmodule Socketfight.GameState do
   def tick() do
     Enum.each(players(), fn {_key, player} ->
       # Setup shoot cooldown
-      player =
-        cond do
-          player.state.shootCooldown > 0 ->
-            update_in(player, [:state, :shootCooldown], fn shootCooldown -> shootCooldown - 1 end)
-
-          true ->
-            player
-        end
+      player = Player.handle_shoot_cooldown(player)
 
       # Reset collision state.
-      updated_player = update_in(player, [:state, :collision], fn _ -> false end)
+      updated_player = Player.reset_collision(player)
 
       # Reset shot state.
-      updated_player = update_in(updated_player, [:state, :shot], fn _ -> false end)
+      updated_player = Player.reset_shot_state(updated_player)
 
       # Filter out names of active actions.
       taken_actions =
@@ -112,15 +106,21 @@ defmodule Socketfight.GameState do
       # Run action handlers.
       updated_player =
         taken_actions
-        |> Enum.reduce(updated_player, fn action, player -> handle_player(player, action) end)
+        |> Enum.reduce(updated_player, fn action, player ->
+          Player.handle_action(player, action)
+        end)
 
       # Handle damage dealt by shooting.
-      deal_damage(updated_player, players(), updated_player.state.shot)
+        damaged_players = Player.deal_damage(updated_player, players(), updated_player.state.shot)
+
+        if is_list(damaged_players) do
+          Enum.each(damaged_players, fn(x) -> update_player(x) end)
+        end
 
       # Clean players with less than zero health.
       players()
       |> Enum.filter(fn {_, player} -> player.state.health <= 0 end)
-      |> Enum.map(fn {_, player} -> handle_death(player) end)
+      |> Enum.map(fn {_, player} -> Player.handle_action(player, "death") |> update_player() end)
 
       # Run collision detection. If no collisions, move player. Otherwise cancel move.
       if !Enum.any?(obstacles(), fn obstacle ->
@@ -145,108 +145,11 @@ defmodule Socketfight.GameState do
           update_in(updated_player, [:state, :y], fn _ -> updated_player.state.newY end)
 
         # Update player.
-        updated_player |> update_player
+        updated_player |> update_player()
       else
         # Update player.
-        updated_player |> update_player
+        updated_player |> update_player()
       end
     end)
-  end
-
-  @doc """
-  Check if player has shot and deal damage to players hit by bullet. Returns
-  a player with updated kills status.
-  """
-  def deal_damage(player, players, true) do
-    # Loop through each opponent. Check collisions.
-    players_hit =
-      players
-      |> Enum.filter(fn {_, target_player} ->
-        elem(
-          CollisionDetector.collides?(
-            target_player.state.x,
-            target_player.state.y,
-            target_player.radius,
-            player.state.x,
-            player.state.y,
-            player.state.shootTargetX,
-            player.state.shootTargetY
-          ),
-          0
-        )
-      end)
-
-    # Order list based on distance.
-
-    # Loop through hit players
-    players_hit
-    |> Enum.map(fn {_, target_player} ->
-      # Reduce target player health.
-      update_in(target_player, [:state, :health], fn health -> health - 20 end)
-      |> update_player
-    end)
-  end
-
-  @doc """
-  In case weapon is not shot, no damage is done.
-  """
-  def deal_damage(player, _, _) do
-    player
-  end
-
-  @doc """
-  If player dies, increase death counter and reset location.
-  """
-  def handle_death(player) do
-    player
-    |> update_in([:state, :health], fn _ -> 100 end)
-    |> update_in([:state, :deaths], fn deaths -> deaths + 1 end)
-    |> update_player
-  end
-
-  # def handle_bullet_damage(player, players) do
-  # end
-
-  def handle_player(player, "forward") do
-    xOffset = :math.cos(player.state.rotation + :math.pi() / 2) * 5
-    yOffset = :math.sin(player.state.rotation + :math.pi() / 2) * 5
-    player = update_in(player, [:state, :newX], fn _ -> player.state.x - xOffset end)
-    update_in(player, [:state, :newY], fn _ -> player.state.y - yOffset end)
-  end
-
-  def handle_player(player, "brake") do
-    xOffset = :math.cos(player.state.rotation + :math.pi() / 2) * 5
-    yOffset = :math.sin(player.state.rotation + :math.pi() / 2) * 5
-    player = update_in(player, [:state, :newX], fn x -> x + xOffset end)
-    update_in(player, [:state, :newY], fn y -> y + yOffset end)
-  end
-
-  def handle_player(player, "left") do
-    update_in(player, [:state, :rotation], fn rotation -> rotation - :math.pi() / 60 end)
-  end
-
-  def handle_player(player, "right") do
-    update_in(player, [:state, :rotation], fn rotation -> rotation + :math.pi() / 60 end)
-  end
-
-  @doc """
-  Shoot weapon.
-  """
-  def handle_player(player, "shoot") do
-    # If cooldown is over and shoot-button is enabled.
-    if player.state.shootCooldown == 0 && player.actions["shoot"] == true do
-      # Calculate end point.
-      xOffset = :math.cos(player.state.rotation + :math.pi() / 2) * 400
-      yOffset = :math.sin(player.state.rotation + :math.pi() / 2) * 400
-      IO.puts("Shoot to: #{xOffset} #{yOffset}")
-
-      # Update states.
-      player = update_in(player, [:state, :shootTargetX], fn _ -> player.state.x - xOffset end)
-      player = update_in(player, [:state, :shootTargetY], fn _ -> player.state.y - yOffset end)
-      player = update_in(player, [:state, :shot], fn _ -> true end)
-      update_in(player, [:state, :shootCooldown], fn _ -> 30 end)
-    else
-      player
-    end
   end
 end
